Webpack-react example now works
