+++ .yarnrc.yml
@@ -1,1 +1,3 @@
 yarnPath: .yarn/releases/yarn-3.6.3.cjs
+nodeLinker: pnp
+pnpMode: loose
