+++ examples/_webpack-config/.babelrc
@@ -1,0 +1,4 @@
+{
+  "presets": ["@babel/preset-env", "@babel/preset-typescript"]
+}
+
