+++ examples/_webpack-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "example-webpack-react",
+  "name": "webpack-config",
   "private": true,
   "scripts": {
     "dev": "webpack-dev-server ./src/index.ts"
@@ -7,12 +7,8 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@macrostrat/hyper": "^2.2.1",
     "cesium": "^1.109.0",
-    "core-js": "^3.19.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "resium": "^1.13.1"
+    "core-js": "^3.19.0"
   },
   "devDependencies": {
     "@babel/core": "^7.19.6",
@@ -30,7 +26,6 @@
     "typescript": "^5",
     "webpack": "^5.64.1",
     "webpack-cli": "^4.9.1",
-    "webpack-config": "workspace:*",
     "webpack-dev-server": "^4.3.1"
   }
 }
