+++ examples/webpack/package.json
@@ -26,6 +26,7 @@
     "typescript": "^5",
     "webpack": "^5.64.1",
     "webpack-cli": "^4.9.1",
+    "webpack-config": "workspace:*",
     "webpack-dev-server": "^4.3.1"
   }
 }
